+++ src/projects/forms/frm_project_edit.py
@@ -6,6 +6,7 @@ import tkinter as tk
 from pathlib import Path
 from tkinter import colorchooser, filedialog, messagebox, ttk
 
+from projects.data_store import store as data_store
 from psiutils.buttons import ButtonFrame, IconButton
 from psiutils.constants import PAD, Mode, Status, WidgetState
 from psiutils.utilities import geometry, window_resize
@@ -15,7 +16,6 @@ from projects import logger
 from projects.config import config
 from projects.constants import APP_TITLE, ICON_DIR
 from projects.project import Project
-from projects.project_store import store as project_store
 from projects.text import Text
 
 txt = Text()
@@ -416,7 +416,7 @@ class ProjectEditFrame:
             self.project.workbench_colours[key] = colour
 
         logger.info("Project changed", changes=changes)
-        project_store.save_projects(self.projects)
+        data_store.save_projects(self.projects)
         self.project_version.set(self.project.version_text)
         self.status = Status.UPDATED
         self._dismiss()